+++ Libplanet.Tests/Net/SwarmTest.cs
@@ -2017,13 +2017,13 @@ namespace Libplanet.Tests.Net
         }
 
         [Fact(Timeout = Timeout)]
-        public async Task GetPeerChainStatus()
+        public async Task GetPeerChainStatusAsync()
         {
             Swarm<DumbAction> swarm1 = _swarms[0];
             Swarm<DumbAction> swarm2 = _swarms[1];
             Swarm<DumbAction> swarm3 = _swarms[2];
 
-            var peerStatus = await swarm1.GetPeerChainStatus(TimeSpan.FromSeconds(1), default);
+            var peerStatus = await swarm1.GetPeerChainStatusAsync(TimeSpan.FromSeconds(1), default);
             Assert.Empty(peerStatus);
 
             try
@@ -2033,21 +2033,21 @@ namespace Libplanet.Tests.Net
 
                 await BootstrapAsync(swarm1, swarm2.AsPeer);
 
-                peerStatus = await swarm1.GetPeerChainStatus(TimeSpan.FromSeconds(1), default);
+                peerStatus = await swarm1.GetPeerChainStatusAsync(TimeSpan.FromSeconds(1), default);
                 Assert.Equal(
                     new PeerChainStatus((BoundPeer)swarm2.AsPeer, 0, 0),
                     peerStatus.First()
                 );
 
                 await swarm2.BlockChain.MineBlock(_fx1.Address1);
-                peerStatus = await swarm1.GetPeerChainStatus(TimeSpan.FromSeconds(1), default);
+                peerStatus = await swarm1.GetPeerChainStatusAsync(TimeSpan.FromSeconds(1), default);
                 Assert.Equal(
                     new PeerChainStatus((BoundPeer)swarm2.AsPeer, 1, 1024),
                     peerStatus.First()
                 );
 
                 await BootstrapAsync(swarm1, swarm3.AsPeer);
-                peerStatus = await swarm1.GetPeerChainStatus(TimeSpan.FromSeconds(1), default);
+                peerStatus = await swarm1.GetPeerChainStatusAsync(TimeSpan.FromSeconds(1), default);
                 Assert.Equal(
                     new[]
                     {